+++ src/nsidc/iceflow/api.py
@@ -1,55 +1,21 @@
+"""Module with functions defining the user-facing API for iceflow."""
+
 from __future__ import annotations
 
-import datetime as dt
 import shutil
 from pathlib import Path
 
 import dask.dataframe as dd
-import pandas as pd
 from loguru import logger
 
-from nsidc.iceflow.data.fetch import search_and_download
+from nsidc.iceflow.data.fetch import download_iceflow_results, find_iceflow_data
 from nsidc.iceflow.data.models import (
-    BoundingBox,
-    Dataset,
+    ALL_DATASETS,
     DatasetSearchParameters,
     IceflowDataFrame,
 )
-from nsidc.iceflow.data.read import read_data
+from nsidc.iceflow.data.read import read_iceflow_datafiles
 from nsidc.iceflow.itrf.converter import transform_itrf
-
-
-def _df_for_one_dataset(
-    *,
-    dataset: Dataset,
-    bounding_box: BoundingBox,
-    temporal: tuple[dt.datetime | dt.date, dt.datetime | dt.date],
-    output_dir: Path,
-    # TODO: also add option for target epoch!!
-    output_itrf: str | None,
-) -> IceflowDataFrame:
-    results = search_and_download(
-        short_name=dataset.short_name,
-        version=dataset.version,
-        bounding_box=bounding_box,
-        temporal=temporal,
-        output_dir=output_dir,
-    )
-
-    all_dfs = []
-    for result in results:
-        data_df = read_data(dataset, result)
-        all_dfs.append(data_df)
-
-    complete_df = IceflowDataFrame(pd.concat(all_dfs))
-
-    if output_itrf is not None:
-        complete_df = transform_itrf(
-            data=complete_df,
-            target_itrf=output_itrf,
-        )
-
-    return complete_df
 
 
 def fetch_iceflow_df(
@@ -67,38 +33,43 @@ def fetch_iceflow_df(
     user requests a large spatial/temporal area across multiple datasets. The
     result may not even fit in memory!
 
-    Consider using `create_iceflow_parquet` to fetch and store data in parquet
+    Consider using `make_iceflow_parquet` to store downloaded data in parquet
     format.
     """
 
-    dfs = []
-    for dataset in dataset_search_params.datasets:
-        result = _df_for_one_dataset(
-            dataset=dataset,
-            temporal=dataset_search_params.temporal,
-            bounding_box=dataset_search_params.bounding_box,
-            output_dir=output_dir,
-            output_itrf=output_itrf,
+    iceflow_search_reuslts = find_iceflow_data(
+        dataset_search_params=dataset_search_params,
+    )
+
+    downloaded_files = download_iceflow_results(
+        iceflow_search_results=iceflow_search_reuslts,
+        output_dir=output_dir,
+    )
+
+    iceflow_df = read_iceflow_datafiles(downloaded_files)
+
+    if output_itrf is not None:
+        iceflow_df = transform_itrf(
+            data=iceflow_df,
+            target_itrf=output_itrf,
         )
-        dfs.append(result)
 
-    complete_df = IceflowDataFrame(pd.concat(dfs))
-
-    return complete_df
+    return iceflow_df
 
 
-def create_iceflow_parquet(
+def make_iceflow_parquet(
     *,
-    dataset_search_params: DatasetSearchParameters,
-    output_dir: Path,
+    data_dir: Path,
     target_itrf: str,
     overwrite: bool = False,
     target_epoch: str | None = None,
 ) -> Path:
-    """Create a parquet dataset containing the lat/lon/elev data matching the dataset search params.
+    """Create a parquet dataset containing the lat/lon/elev data in `data_dir`.
 
-    This function creates a parquet dataset that can be easily used alongside dask,
-    containing lat/lon/elev data.
+    This function creates a parquet dataset that can be easily used alongside
+    dask, containing lat/lon/elev data. Users who are interested in the full
+    data record with all fields provided by data in the `data_dir` should use
+    `read_iceflow_datafiles`.
 
     Note: this function writes a single `iceflow.parquet` to the output
     dir. This code does not currently support updates to the parquet after being
@@ -108,52 +79,47 @@ def create_iceflow_parquet(
     need to move/remove the existing `iceflow.parquet` first (e.g., with the
     `overwrite=True` kwarg).
     """
-    output_subdir = output_dir / "iceflow.parquet"
-    if output_subdir.exists():
+    parquet_subdir = data_dir / "iceflow.parquet"
+    if parquet_subdir.exists():
         if overwrite:
             logger.info("Removing existing iceflow.parquet")
-            shutil.rmtree(output_subdir)
+            shutil.rmtree(parquet_subdir)
         else:
             raise RuntimeError(
                 "An iceflow parquet file already exists. Use `overwrite=True` to overwrite."
             )
 
-    for dataset in dataset_search_params.datasets:
-        results = search_and_download(
-            short_name=dataset.short_name,
-            version=dataset.version,
-            temporal=dataset_search_params.temporal,
-            bounding_box=dataset_search_params.bounding_box,
-            output_dir=output_dir,
+    all_subdirs = [
+        data_dir / ds.subdir_name
+        for ds in ALL_DATASETS
+        if (data_dir / ds.subdir_name).is_dir()
+    ]
+    for subdir in all_subdirs:
+        iceflow_filepaths = [path for path in subdir.iterdir() if path.is_file()]
+        iceflow_df = read_iceflow_datafiles(iceflow_filepaths)
+
+        iceflow_df = transform_itrf(
+            data=iceflow_df,
+            target_itrf=target_itrf,
+            target_epoch=target_epoch,
         )
 
-        for result in results:
-            data_df = read_data(dataset, result)
-            df = IceflowDataFrame(data_df)
-
-            df = transform_itrf(
-                data=df,
-                target_itrf=target_itrf,
-                target_epoch=target_epoch,
+        # Add a string col w/ dataset name and version.
+        short_name, version = subdir.name.split("_")
+        iceflow_df["dataset"] = [f"{short_name}v{version}"] * len(iceflow_df.latitude)
+        common_columns = ["latitude", "longitude", "elevation", "dataset"]
+        common_dask_df = dd.from_pandas(iceflow_df[common_columns])  # type: ignore[attr-defined]
+        if parquet_subdir.exists():
+            dd.to_parquet(  # type: ignore[attr-defined]
+                df=common_dask_df,
+                path=parquet_subdir,
+                append=True,
+                ignore_divisions=True,
+            )
+        else:
+            dd.to_parquet(  # type: ignore[attr-defined]
+                df=common_dask_df,
+                path=parquet_subdir,
             )
 
-            # Add a string col w/ dataset name and version.
-            df["dataset"] = [f"{dataset.short_name}v{dataset.version}"] * len(
-                df.latitude
-            )
-            common_columns = ["latitude", "longitude", "elevation", "dataset"]
-            common_dask_df = dd.from_pandas(df[common_columns])  # type: ignore[attr-defined]
-            if output_subdir.exists():
-                dd.to_parquet(  # type: ignore[attr-defined]
-                    df=common_dask_df,
-                    path=output_subdir,
-                    append=True,
-                    ignore_divisions=True,
-                )
-            else:
-                dd.to_parquet(  # type: ignore[attr-defined]
-                    df=common_dask_df,
-                    path=output_subdir,
-                )
-
-    return output_subdir
+    return parquet_subdir