+++ tasks/docs.py
@@ -50,9 +50,9 @@ def watch(_ctx):
 
 
 @task()
-def generate_example_notebook(_ctx):
-    """Generate the example iceflow notebook for the docs."""
+def render_notebooks_for_docs(_ctx):
+    """Render jupyter notebooks for docs."""
     print_and_run(
-        (f"{PROJECT_DIR}/notebooks/generate_rendered_notebook_for_docs.sh"),
+        (f"{PROJECT_DIR}/notebooks/generate_rendered_notebooks_for_docs.sh"),
         pty=True,
     )