+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add support for ILVIS2 v1 and v2.
 - Improve API, providing ability to search across datasets and save to
   intermediate parquet file.
+- Add jupyter notebook showing use of `iceflow` alongside `icepyx`
 
 # v0.2.0