+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   `target_epoch`.
 - Add support for ILATM1B v2 and BLATM1B v1.
 - Add support for ILVIS2 v1 and v2.
+- Improve API, providing ability to search across datasets and save to
+  intermediate parquet file.
+- Add jupyter notebook showing use of `iceflow` alongside `icepyx`
 
 # v0.2.0